refactor: move PrintBlockTree to node/blockstorage; delete main.cpp (#3125 C9)
PrintBlockTree (the -printblocktree debug dump, called from AppInit2)
joins the rest of the block-file machinery in node/blockstorage.{h,cpp};
blockstorage.h gains the sync.h + extern-cs_main idiom for the lock
annotation, and main.h loses the declaration (blockstorage.h is part of
its re-include set).

That was the last definition in main.cpp: delete the file and its build
registration. main.h remains as a compatibility umbrella -- re-includes,
a small fenced block of duplicate declarations, and breadcrumb comments
-- until the core TUs that still reach the duplicated declarations only
through it are repointed (tracked in #3125). No src/qt TU is among them:
after the Phase-1e/1f extraction (#3204, #3208) src/qt includes main.h
nowhere, and lint-qt-includes.sh now hard-fails any direct core-header
include outside the composition root, so main.h's remaining consumers are
core-side only.

Pure code motion; no behavior change.
